Rectangle Project File
Uploading files to branch. Rectangle defines a rectangle by holding two points and has two methods to calculate the area of the rectangle and the length of its diagonal. It creates two rectangles with different points and checks  to see if the area and diagonal matches the expected values.

There are bugs in the code that cause the JUnit test to fail. Find the bugs in the code to improve the quality of the code as well as allow the JUnit test to pass.