Fibonacci Bug Fix
There was a bug in Fibonacci.java that caused JUnit test to fail. The bug was that when the test ran and Fibonacci used 0 the fail JUnit test said that it expected 0 but was 1. In the class we needed to change the case 0 from returning 1 to 0. There were two cases where 1 would be returned for both 0 and 1.